requirements-fixed.txt
flask==2.0.1
flask-cors==3.0.10
openai-whisper==20230918
pydub==0.25.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.1+cpu
torchvision==0.16.1+cpu
torchaudio==2.1.1+cpu
transformers==4.30.2
pymongo==4.3.3
numpy==1.24.3
youtube-dl==2021.12.17
yt-dlp==2023.3.4
python-dotenv==1.0.0
